chatbot/requirements.txt
flask
pandas
nltk
scikit-learn
pyspellchecker
